refactor(FE): Gyms page
